+++ Projects/Features/PlayerFeature/Sources/ViewModels/PlayerViewModel.swift
@@ -64,8 +64,8 @@ final class PlayerViewModel: ViewModelType {
     var fetchLikeNumOfSongUseCase: FetchLikeNumOfSongUseCase!
     var fetchFavoriteSongsUseCase: FetchFavoriteSongsUseCase!
     
+    let disposeBag = DisposeBag()
     private let playState = PlayState.shared
-    private let disposeBag = DisposeBag()
     private var subscription = Set<AnyCancellable>()
     internal var lyricsDict = [Float : String]()
     internal var sortedLyrics = [String]()
@@ -283,147 +283,4 @@ final class PlayerViewModel: ViewModelType {
         output.totalTimeValue.send(Float(progress.endProgress))
     }
     
-    /// 가사 불러오기
-    func fetchLyrics(for song: SongEntity, output: Output) {
-        fetchLyricsUseCase.execute(id: song.id)
-            .retry(3)
-            .subscribe { [weak self] lyricsEntityArray in
-                guard let self else { return }
-                self.lyricsDict.removeAll()
-                self.sortedLyrics.removeAll()
-                lyricsEntityArray.forEach { self.lyricsDict.updateValue($0.text, forKey: Float($0.start)) }
-                self.sortedLyrics = self.lyricsDict.sorted { $0.key < $1.key }.map { $0.value }
-            } onFailure: { [weak self] error in
-                guard let self else { return }
-                self.lyricsDict.removeAll()
-                self.sortedLyrics.removeAll()
-                self.sortedLyrics.append("가사가 없습니다.")
-                print("title: \(song.title) id: \(song.id) 가사가 없습니다. error: \(error)")
-            } onDisposed: {
-                output.lyricsDidChangedEvent.send(true)
-            }.disposed(by: self.disposeBag)
-    }
-    
-    /// 좋아요 수 가져오기
-    func fetchLikeCount(for song: SongEntity, output: Output) {
-        fetchLikeNumOfSongUseCase.execute(id: song.id)
-            .retry(3)
-            .map { [weak self] song in
-                self?.formatNumber(song.likes) ?? ""
-            }
-            .subscribe { likeCountText in
-                output.likeCountText.send(likeCountText)
-            } onFailure: { _ in
-                output.likeCountText.send("좋아요")
-            }.disposed(by: self.disposeBag)
-    }
-    
-    /// 좋아요를 누른 곡인지 여부 판별
-    func fetchLikeState(for song: SongEntity, output: Output) {
-        fetchFavoriteSongsUseCase.execute()
-            .retry(3)
-            .map { $0.contains { $0.song.id == song.id } }
-            .subscribe { isLiked in
-                output.likeState.send(isLiked)
-            } onFailure: { _ in
-                output.likeState.send(false)
-            }.disposed(by: self.disposeBag)
-    }
-    
-    func cancelLikeSong(for song: SongEntity, output: Output) {
-        self.cancelLikeSongUseCase.execute(id: song.id)
-            .retry(3)
-            .subscribe(
-                onSuccess: { _ in
-                    output.likeState.send(false)
-                },
-                onFailure: { error in
-                    output.showToastMessage.send(error.localizedDescription)
-                }
-            )
-            .disposed(by: self.disposeBag)
-    }
-    
-    func addLikeSong(for song: SongEntity, output: Output) {
-        self.addLikeSongUseCase.execute(id: song.id)
-            .retry(3)
-            .subscribe(
-                onSuccess: { _ in
-                    output.likeState.send(true)
-                },
-                onFailure: { error in
-                    output.showToastMessage.send(error.localizedDescription)
-                }
-            )
-            .disposed(by: self.disposeBag)
-    }
-    
-}
-
-// MARK: - 뷰모델 내 유틸리티 함수들을 모아놓은 곳입니다.
-extension PlayerViewModel {
-    func formatTime(_ second: Double) -> String {
-        let second = Int(floor(second))
-        let min = second / 60
-        let sec = String(format: "%02d", second % 60)
-        return "\(min):\(sec)"
-    }
-    
-    func formatNumber(_ number: Int) -> String {
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .decimal
-        formatter.groupingSeparator = ""
-        formatter.decimalSeparator = "."
-        formatter.maximumFractionDigits = 2
-        formatter.minimumFractionDigits = 0
-        
-        switch number {
-        case 0..<1000:
-            return String(number)
-        case 1000..<10_000:
-            let thousands = Double(number) / 1000.0
-            return formatter.string(from: NSNumber(value: thousands))! + "천"
-        case 10_000..<100_000_0:
-            let tenThousands = Double(number) / 10000.0
-            return formatter.string(from: NSNumber(value: tenThousands))! + "만"
-        case 100_000_0..<100_000_000:
-            let tenThousands = Int(number) / 10000
-            return formatter.string(from: NSNumber(value: tenThousands))! + "만"
-        default:
-            let millions = Double(number) / 100000000.0
-            return formatter.string(from: NSNumber(value: millions))! + "억"
-        }
-    }
-    
-    func getCurrentLyricsIndex(_ currentTime: Float) -> Int {
-        let times = lyricsDict.keys.sorted()
-        let index = closestIndex(to: currentTime, in: times)
-        return index
-    }
-    
-    /// target보다 작거나 같은 값 중에서 가장 가까운 값을 찾습니다. O(log n)
-    func closestIndex(to target: Float, in arr: [Float]) -> Int {
-        var left = 0
-        var right = arr.count - 1
-        var closestIndex = 0
-        var closestDistance = Float.greatestFiniteMagnitude
-        
-        while left <= right {
-            let mid = (left + right) / 2
-            let midValue = arr[mid]
-            
-            if midValue <= target {
-                let distance = target - midValue
-                if distance < closestDistance {
-                    closestDistance = distance
-                    closestIndex = mid
-                }
-                left = mid + 1
-            } else {
-                right = mid - 1
-            }
-        }
-        
-        return closestIndex
-    }
 }
